Comments and assertions for @koadman code
AD-- it's not clear to me that this is doing the right thing. For
example, it does not seem that we will get the posterior distribution if
we, say, run n sequences until they make a tree, then add m sequences
and start the online algorithm. For example, imagine that there is no
data and n=m are both quite large. We won't get the uniform distribution
on cherries (2-leaf subtrees) because it will be much more likely for
the m leaves to be in a cherry than one of the m sequences and one of
the n sequences. Does that seem right?